Revert "Abstract create table"
This reverts commit 22f170db7400406f32687b98caad7c2810123060.
